+++ src/app.ts
@@ -1,10 +1,9 @@
 require('dotenv').config('../.env')
 import 'module-alias/register';
 import * as express from 'express';
-import { projectRouter, authenticationRouter } from './routes';
-import {createContext, purgeModel} from '@utilities';
+import { projectRouter, authenticationRouter, userRouter } from './routes';
+import {createContext} from '@utilities';
 import { errorHandler, logError, authenticate } from '@middlewares';
-import { createRolesAndPermissionsAndResources } from '@dbmethods/authorization';
 import { debugRouter  } from './routes/debug';
 
 // Connects to MongoDB Atlas
@@ -16,10 +15,11 @@ const port:number = 3000;
 
 app.use(express.json());
 app.all('*', createContext);
-app.use("/debug", debugRouter()); //===== DEBUG DEBUG DEBUG 
+// uncomment for debugging app.use("/debug", debugRouter()); 
 app.use("/user", authenticationRouter());
 app.all('/api/*', authenticate)
 app.use("/api/project", projectRouter(1));
+app.use("/api/user", userRouter(1)); 
 app.use(logError);
 app.use(errorHandler);
